Remove old header and footer, my listings page and new listing page
