+++ server/server.js
@@ -167,7 +167,7 @@ const startServer = () => {
       })
 
       server.get('/meus-imoveis', (req, res) => {
-        return app.render(req, res, '/listings/user-listings', req.query)
+        return app.render(req, res, '/user/profile', req.query)
       })
 
       server.get('/busca', (req, res) => {
